+++ redisson-spring-data/redisson-spring-data-23/src/main/java/org/redisson/spring/data/connection/RedissonConnection.java
@@ -1872,7 +1872,7 @@ public class RedissonConnection extends AbstractRedisConnection {
         if (returnType == ReturnType.BOOLEAN) {
             c = org.redisson.api.RScript.ReturnType.BOOLEAN.getCommand();
         } else if (returnType == ReturnType.INTEGER) {
-            c = org.redisson.api.RScript.ReturnType.INTEGER.getCommand();
+            c = org.redisson.api.RScript.ReturnType.LONG.getCommand();
         } else if (returnType == ReturnType.MULTI) {
             c = org.redisson.api.RScript.ReturnType.LIST.getCommand();
             return new RedisCommand(c, name, new BinaryConvertor());
